Split the market page into components, fix title
