+++ app/(root)/market/market.module.scss
@@ -1,150 +1,8 @@
 .mainContainer {
-  // width: 71.5vw;
   width: 100%;
   max-height: 100%;
   min-height: 70vh;
   padding-right: 65px;
+  overflow: hidden;
+  
 }
-
-.container {
-  display: flex;
-  justify-content: start;
-  align-items: start;
-  flex-wrap: wrap;
-  gap: 10px;
-  padding: 20px;
-  width: 100%;
-  max-height: 100vh;
-  min-height: 70vh;
-  background-color: #3a4340;
-  border-radius: 10px;
-  overflow-y: scroll;
-  position: relative;
-}
-
-.card {
-  border-radius: 10px;
-  width: 179px;
-  height: 222px;
-  box-shadow: 0 5px 15px 0 rgba(32, 36, 34, 0.8);
-  background: #232927;
-  display: flex;
-  flex-direction: column;
-  padding: 10px;
-  color: #d9d9d9;
-}
-
-.productImg {
-  width: 100%;
-  height: 126px;
-  object-fit: cover;
-  margin-bottom: 2px;
-  border-radius: 5px;
-}
-
-.title {
-  font-weight: 700;
-  font-size: 16px;
-  line-height: 18px;
-}
-
-.price {
-  font-weight: 500;
-  font-size: 14px;
-  display: flex;
-  justify-content: end;
-  align-items: end;
-  margin-bottom: 2px;
-  span {
-    margin-right: 5px;
-  }
-  img {
-    margin-bottom: 3px;
-  }
-}
-
-.btnBorder {
-  width: 100%;
-  height: 31px;
-  padding: 1px;
-  background: #474C52;
-  border-radius: 0 10px;
-  .buying {
-    width: 100%;
-    height: 100%;
-    background: #181819;
-    border-radius: 0 10px;
-    display: flex;
-    align-items: center;
-
-    img {
-      margin: 0 11px 0 10px;
-    }
-
-    p {
-      font-weight: 400;
-      font-size: 14px;
-    }
-
-    &:active {
-      background: #0F629A;
-    }
-  }
-  &:hover {
-    background: linear-gradient(90deg, rgba(24, 134, 208, 1) 0%, rgba(12, 68, 106, 1) 100%);
-  }
-}
-
-
-
-.botCard {
-  flex: 15%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.seller {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  border-radius: 2px;
-  background-color: rgba(47, 47, 47, 1);
-  width: 60px;
-}
-
-.count {
-  img {
-    float: left;
-    width: 20px;
-    height: 20px;
-  }
-  p {
-    font-size: 12px;
-  }
-}
-
-.link {
-  display: flex;
-  border-left: 1px solid rgba(0, 0, 0, 1);
-}
-
-
-.container::-webkit-scrollbar {
-  width: 8px;
-  position: absolute;
-  left: 50px;
-}
-
-.container::-webkit-scrollbar-track {
-  background-color: transparent;
-  border-radius: 8px;
-}
-
-.container::-webkit-scrollbar-thumb {
-  background-color: #D9D9D9;;
-  border-radius: 20px;
-  background-clip: padding-box;
-}
-
-
