+++ app/components/ProductCard/index.tsx
@@ -1,13 +1,44 @@
-import React from 'react'
-import styles from './card.module.scss'
-import Image from "next/image";
+"use client"
 
-const index = () => {
+import React, { useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
+import Image from 'next/image';
+import styles from './card.module.scss';
+import { Link } from 'lucide-react';
+import { IProduct } from '@/index';
+import iconCart from '../../../public/Icons-basket.svg';
+import iconPrice from '../../../public/icons-price.svg';
+
+const ProductCard: React.FC<{ product: IProduct }> = ({ product }) => {
   return (
-    <div>
-        Card
-    </div>
-  )
-}
+    <div className={styles.card}>
+      <div>
+        <Image
+          className={styles.productImg}
+          src={product.image}
+          alt={product.alt}
+          width={200}
+          height={200}
+        />
+      </div>
 
-export default index
+      <div className={styles.title}>
+        <h3>{product.name}</h3>
+      </div>
+
+      <div className={styles.price}>
+        <span>{product.price}</span>
+        <Image src={iconPrice} alt="iconPrice" width={20} height={20} />
+      </div>
+
+      <div className={styles.btnBorder}>
+        <button type="button" className={styles.buying}>
+          <Image src={iconCart} alt="icon-basket" width={29} height={29} />
+          <p>Add to cart</p>
+        </button>
+      </div>
+    </div>
+  );
+};
+
+export default ProductCard;
