+++ app/components/ProductsBox/fakeData.ts
